Started work on options dialog updates:
Before, NVDA Remote users had the ability to automatically connect to a remote control server when NVDA started up. It was not possible to connect to a server hosted on the same machine automatically.
with these changes in dialogs.py, the user can now choose to have their copy of NVDA Remote start it's own local server, ready to be directly connected to and controlled, verses having their addon connect to a potentially compromised remote server they might not have full trust in.