First stab at handling nvdaremote:// links
This uses a tiny C++ app which you can build with build_url_handler.bat assuming you're in a Visual Studio command prompt.
The C++ app takes its command line and squirts it over using the WM_COPYDATA message.
The actual NVDA side receives this message, decodes the URL out of it, and prompts the user.
@tspivey wrote the C++ portion.