Block the main thread when connecting on the secure desktop so we don't miss the first output.
Before this, when switching to a secure desktop, you'd often (always?) miss the first output from NVDA, usually the UAC message.
This could be a little confusing.
This occurred because the connection took a short time to establish, but by then, NVDA had already reported the initial focus.
To get around this, block the main thread until the connection is established.
Since we're connecting to localhost, this should be pretty fast, but we use a 1 second timeout just in case.
This is a bit ugly, but I think it's cleaner and less risky than the alternative of queuing the output until the connection is established.